@@ -5,15 +5,17 @@ import android.animation.ObjectAnimator;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.provider.ContactsContract;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
@@ -138,12 +140,15 @@ public class EndGame extends AppCompatActivity {
         }
     }
     private void requestPost() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(this).setMessage(R.string.post_message).setPositiveButton("Post", new DialogInterface.OnClickListener() {
+        final AlertDialog.Builder builder = new AlertDialog.Builder(this).setMessage(R.string.post_message).setPositiveButton("Post", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialogInterface, int i) {
                 if (!posted) {
                     saveScoresDatabase();
                     saveScoresLocal();
+                }
+                if (savedScores) {
+                    Toast.makeText(getBaseContext(),"Scores Already Saved", Toast.LENGTH_LONG).show();
                 }
             }
         }).setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -172,6 +177,7 @@ public class EndGame extends AppCompatActivity {
                     //databaseUtils.getDatabaseInstance().getReference().child("Users").child(databaseUtils.getUserID()).child("Past Exams").child();
                     //databaseUtils.getDatabaseInstance().getReference().child("Users").child(databaseUtils.getUserID()).child("PastExams").child()
                     savedScores = true;
+                    Toast.makeText(getBaseContext(), "Score Has Been Saved", Toast.LENGTH_LONG).show();
                 }
             }
 
@@ -180,6 +186,7 @@ public class EndGame extends AppCompatActivity {
 
             }
         });
+
     }
 
     private void saveScoresLocal() {
