@@ -1,7 +1,8 @@
 package com.corp.srihari.deca;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class Learn extends AppCompatActivity {
 
